+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Pending
 
+* [Load V5] Fix Event Protocol and Cleanup [#16560](https://github.com/Applifier/unity-ads-webview/pull/16560)
+* [Load V5] Send load api enabled flag to kafka [#16434](https://github.com/Applifier/unity-ads-webview/pull/16434)
+* Add load and fill event manager [#16401](https://github.com/Applifier/unity-ads-webview/pull/16401)
+
+# Tue Jul 21 21:43:41 UTC 2020
+
 * [VAST Endcard] Lock Screen Orientation on Show [#16495](https://github.com/Applifier/unity-ads-webview/pull/16495)
 * [AUI] Make MAB reverse AB test explicit [#16436](https://github.com/Applifier/unity-ads-webview/pull/16436)
 * [VAST Endcard PR1] Refactor show method to be async and add more unit tests [#16252](https://github.com/Applifier/unity-ads-webview/pull/16252)